+++ software/src/heston_ml_both.hpp
@@ -38,7 +38,7 @@ private:
 	uint32_t price_cnt;
 };
 
-float heston_ml_control(const HestonParamsML &ml_params,
+double heston_ml_control(const HestonParamsML &ml_params,
 		std::function<Statistics(const HestonParamsML, const uint32_t, 
 		const uint64_t, const bool, const uint32_t)> ml_kernel);
